requirements.txt
# Solventación - Sistema de Procesamiento de Documentos
# Dependencias del proyecto

# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Procesamiento de documentos
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.3.3

# Base de datos
python-dotenv==1.0.0

# OpenAI API
openai==1.12.0

# Utilidades
python-multipart==0.0.6
